requirements.txt
# 图像处理
Pillow>=10.0.0
exifread>=3.0.0
imagehash>=4.3.0

# 人脸识别
face_recognition>=1.3.0
dlib>=19.24.0

# vLLM
vllm==0.16.0
transformers>=4.40.0

# Web
flask>=3.0.0

# 数据库
# sqlite3 (内置)

# 工具
tqdm>=4.65.0
